use custom require and requireNotNull
